refactor(catalog): remove interface
