+++ src/main/java/me/helioalbano/biblioteca/catalog/config/CatalogConfig.java
@@ -1,12 +1,10 @@
 package me.helioalbano.biblioteca.catalog.config;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
 import me.helioalbano.biblioteca.catalog.domain.repository.BookRepository;
 import me.helioalbano.biblioteca.catalog.infra.repository.BookRepositoryPostgres;
-import me.helioalbano.biblioteca.catalog.usecase.book.create.Create;
-import me.helioalbano.biblioteca.catalog.usecase.book.create.impl.CreateImpl;
+import me.helioalbano.biblioteca.catalog.usecase.CreateBook;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class CatalogConfig {
@@ -17,7 +15,7 @@ public class CatalogConfig {
     }
 
     @Bean
-    public Create addNewBook() {
-        return new CreateImpl(bookRepository());
+    public CreateBook createBookBook() {
+        return new CreateBook(bookRepository());
     }
 }
