+++ src/main/java/me/helioalbano/biblioteca/catalog/infra/controller/BookController.java
@@ -4,7 +4,7 @@ import java.net.URI;
 
 import javax.validation.Valid;
 
-import org.modelmapper.ModelMapper;
+import me.helioalbano.biblioteca.catalog.usecase.CreateBook;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,20 +12,19 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import me.helioalbano.biblioteca.catalog.infra.controller.dto.CreateBookRequest;
-import me.helioalbano.biblioteca.catalog.usecase.book.create.Create;
 
 @RestController
 @RequestMapping("catalog")
-class CatalogController {
-    private final Create create;
+class BookController {
+    private final CreateBook createBook;
 
-    public CatalogController(Create create) {
-        this.create = create;
+    public BookController(CreateBook createBook) {
+        this.createBook = createBook;
     }
 
     @PostMapping(path = "books")
     public ResponseEntity<Void> createBook(@Valid @RequestBody CreateBookRequest book) {
-        var bookId = create.execute(book.toUseCaseInput());
+        var bookId = createBook.execute(book.toUseCaseInput());
         var location = URI.create("/catalog/books/" + bookId);
         return ResponseEntity.created(location).build();
     }
