--- src/main/java/me/helioalbano/biblioteca/catalog/usecase/book/create/Create.java
@@ -1,7 +1,0 @@
-package me.helioalbano.biblioteca.catalog.usecase.book.create;
-
-import me.helioalbano.biblioteca.catalog.usecase.book.create.dto.CreateBookInput;
-
-public interface Create {
-    Long execute(CreateBookInput createBookInput);
-}
