+++ src/test/java/me/helioalbano/biblioteca/catalog/unit/ModelMapperTest.java
@@ -1,7 +1,7 @@
 package me.helioalbano.biblioteca.catalog.unit;
 
 import me.helioalbano.biblioteca.catalog.infra.controller.dto.CreateBookRequest;
-import me.helioalbano.biblioteca.catalog.usecase.book.create.dto.CreateBookInput;
+import me.helioalbano.biblioteca.catalog.usecase.dto.CreateBookInput;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.modelmapper.ModelMapper;
